refactor: convert js files to ts
Workaround for https://github.com/privatenumber/tsx/issues/38
But also, it's time for those to be TypeScript
Note, I added `// @ts-nocheck` to the top of the files to suppress
TypeScript errors for now, but this should be removed once we move way
from Backbone
